added the two fvec members down() for downcasting to one dimension less clip() for de-homogenization with w-clip
in drawable now exists a new function to compute a ray through a pixel specified by mouse coordinates
ray3 get_world_ray(x,y,view_ptr,modelview_projection_window_matrix)

thick_curve_viewer.cxx now supports gizmo interaction on bezier tube representation with translation and scaling of the control spheres. The new ray computation method is used to allow control sphere picking with Shift-LMB.

+++ cgv/media/color_scale.cxx
@@ -155,9 +155,9 @@ std::vector<float> continuous_color_scale::get_ticks(size_t request_count) const
 	const vec2 domain = get_domain();
 
 	if(mapping_transform_ == ContinuousMappingTransform::Log)
-		return compute_ticks_log(domain[0], domain[1], log_base_, request_count);
+		return compute_ticks_log(domain[0], domain[1], log_base_, (int)request_count);
 	else
-		return compute_ticks(domain[0], domain[1], request_count);
+		return compute_ticks(domain[0], domain[1], (int)request_count);
 }
 
 void continuous_color_scale::set_scheme(const continuous_color_scheme& scheme) {
